% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeCharacter.R
\name{normalizeCharacter}
\alias{normalizeCharacter}
\title{Normalize a Character}
\usage{
normalizeCharacter(
  data,
  character,
  normalize_by,
  new_col_name = paste(character, "_normalized", sep = "")
)
}
\arguments{
\item{data}{A data frame that includes the two columns.}

\item{character}{The name of the column to be normalized (as a string).}

\item{normalize_by}{The name of the column to normalize by (as a string).}

\item{new_col_name}{The name of the new column that will be created (as a string). Default is \code{character} + "_normalized".}
}
\value{
The original data frame with the new column that represents the normalized character.
}
\description{
This function normalizes a morphological character measurement (a column in your tibble)
by another character, following an allometric scaling law.
It fits a power-law relationship (a linear relationship in the log-log space)
between the two features using both a linear model (lm) and a non-linear model (nls).
This methodology is the normalization procedure described in Hamilton et al., 2020 (Journal of Biogeography, 47(7), pp.1494-1509).
}
\examples{
\dontrun{
# Example 1
data <- data.frame(width_cm = rnorm(100, 10, 2), length_cm = rnorm(100, 50, 10))
normalizeCharacter(data, "width_cm", "length_cm")
# Example 2
library(dplyr)
my_tibble <- data.frame(width_cm = rnorm(100, 10, 2), length_cm = rnorm(100, 50, 10))
my_tibble \%>\% normalizeCharacter("width_cm", "length_cm")
}
}
\references{
Hamilton, A.M., Selwyn, J.D., Hamner, R.M., Johnson, H.K., Brown, T., Springer, S.K. & Bird, C.E., 2020.
Biogeography of shell morphology in over‐exploited shellfish reveals adaptive trade‐offs on human‐inhabited
islands and incipient selectively driven lineage bifurcation. Journal of Biogeography, 47(7), pp.1494-1509.
}
